+++ packages/adp-tooling/src/types.ts
@@ -1,6 +1,7 @@
 import type { UI5FlexLayer } from '@sap-ux/project-access';
 import type { DestinationAbapTarget, UrlAbapTarget } from '@sap-ux/system-access';
 import type { Adp } from '@sap-ux/ui5-config';
+import type { Editor } from 'mem-fs-editor';
 
 export interface DescriptorVariant {
     layer: UI5FlexLayer;
@@ -86,6 +87,19 @@ interface CommonChangeProperties {
     texts: Record<string, unknown>;
 }
 
+export interface ManifestChangeProperties {
+    fileName: string;
+    fileType: string;
+    namespace: string;
+    layer: string;
+    packageName: string;
+    reference: string;
+    support: { generator: string };
+    changeType: string;
+    creation: string;
+    content: object;
+}
+
 export interface AddXMLChange extends CommonChangeProperties {
     changeType: 'addXML';
     creation: string;
@@ -113,10 +127,13 @@ export interface CodeExtChange extends CommonChangeProperties {
     };
 }
 
-export const enum FolderNames {
-    Changes = 'changes',
-    Fragments = 'fragments',
-    Coding = 'coding'
+export const enum FolderTypes {
+    CHANGES = 'changes',
+    FRAGMENTS = 'fragments',
+    CODING = 'coding',
+    MANIFEST = 'manifest',
+    ANNOTATIONS = 'annotations',
+    WEBAPP = 'webapp'
 }
 
 export const enum TemplateFileName {
@@ -137,4 +154,196 @@ export const enum HttpStatusCodes {
     INTERNAL_SERVER_ERROR = 500,
     NOT_IMPLEMETED = 501,
     SERVICE_UNAVAILABLE = 503
+}
+
+/**
+ * Represents a constructor type that creates an instance of IWriter.
+ *
+ * @param fs - An instance of Editor used for file system operations.
+ * @param projectPath - The root path of the project.
+ * @returns An instance of IWriter for handling specific data writing operations.
+ */
+export type Writer = new (fs: Editor, projectPath: string) => IWriter<any>;
+
+/**
+ * Generic interface for handling data associated with specific writer operations.
+ * Allows for the typing of data passed to IWriter implementations based on the change type.
+ *
+ * @template T - The subtype of ChangeType that specifies the kind of change and associated data.
+ */
+export type IWriterData<T extends ChangeType> = IWriter<GeneratorData<T>>;
+
+/**
+ * Defines a generic interface for writer classes, specialized by the type of data they handle.
+ *
+ * @template T - The specific type of data the writer will handle, determined by the associated ChangeType.
+ */
+export interface IWriter<T> {
+    /**
+     * Writes the provided data to the project.
+     *
+     * @param data - The data needed for the writer function, specific to the type of change being made.
+     */
+    write(data: T): Promise<void>;
+}
+
+/**
+ * Enumerates the types of changes that can be made, each representing a specific kind of modification.
+ */
+export const enum ChangeType {
+    ADD_NEW_MODEL = 'appdescr_ui5_addNewModel',
+    ADD_ANNOTATIONS_TO_ODATA = 'appdescr_app_addAnnotationsToOData',
+    CHANGE_DATA_SOURCE = 'appdescr_app_changeDataSource',
+    ADD_COMPONENT_USAGES = 'appdescr_ui5_addComponentUsages',
+    ADD_LIBRARY_REFERENCE = 'appdescr_ui5_addLibraries',
+    CHANGE_INBOUND = 'appdescr_app_changeInbound'
+}
+
+/**
+ * Maps a ChangeType to the corresponding data structure needed for that type of change.
+ * This conditional type ensures type safety by linking each change type with its relevant data model.
+ *
+ * @template T - A subtype of ChangeType indicating the specific type of change.
+ */
+export type GeneratorData<T extends ChangeType> = T extends ChangeType.ADD_ANNOTATIONS_TO_ODATA
+    ? AnnotationsData
+    : T extends ChangeType.ADD_COMPONENT_USAGES
+    ? ComponentUsagesData
+    : T extends ChangeType.ADD_LIBRARY_REFERENCE
+    ? ComponentUsagesData
+    : T extends ChangeType.ADD_NEW_MODEL
+    ? NewModelData
+    : T extends ChangeType.CHANGE_DATA_SOURCE
+    ? DataSourceData
+    : T extends ChangeType.CHANGE_INBOUND
+    ? InboundData
+    : never;
+
+export interface AnnotationsData {
+    projectData: AdpProjectData;
+    timestamp: number;
+    /** Indicates whether the annotation is for internal use only. */
+    isInternalUsage: boolean;
+    annotation: {
+        /** Optional name of the annotation file. */
+        fileName?: string;
+        /** Data source associated with the annotation. */
+        dataSource: string;
+        /** Path to the annotation file. */
+        filePath: string;
+    };
+}
+
+export interface ComponentUsagesData {
+    projectData: AdpProjectData;
+    timestamp: number;
+    component: {
+        /** Indicates whether the component is loaded lazily. */
+        isLazy: string;
+        /** Unique ID for the component usage. */
+        usageId: string;
+        /** Name of the component. */
+        name: string;
+        /** Serialized data specific to the component. */
+        data: string;
+        /** Settings related to the component. */
+        settings: string;
+    };
+    library: {
+        /** Reference to the component's library. */
+        reference: string;
+        /** Optional flag indicating if the library reference is lazy. */
+        referenceIsLazy?: string;
+    };
+}
+
+export interface NewModelData {
+    projectData: AdpProjectData;
+    timestamp: number;
+    annotation: {
+        /** Name of the OData annotation data source. */
+        dataSourceName: string;
+        /** Optional URI of the OData annotation data source. */
+        dataSourceURI?: string;
+        /** Optional settings for the OData annotation. */
+        settings?: string;
+    };
+    service: {
+        /** Name of the OData service. */
+        name: string;
+        /** URI of the OData service. */
+        uri: string;
+        /** Name of the OData service model. */
+        modelName: string;
+        /** Version of OData used. */
+        version: string;
+        /** Settings for the OData service model. */
+        modelSettings: string;
+    };
+    /** Indicates whether annotation mode is added. */
+    addAnnotationMode: boolean;
+}
+
+export interface DataSourceData {
+    projectData: AdpProjectData;
+    timestamp: number;
+    service: {
+        /** Data source identifier. */
+        name: string;
+        /** URI of the data source. */
+        uri: string;
+        /** Optional maximum age for the data source cache. */
+        maxAge?: number;
+        /** URI for the OData annotation source. */
+        annotationUri: string;
+    };
+    /** Dictionary mapping data source keys to their values. */
+    dataSourcesDictionary: { [key: string]: string };
+}
+
+export interface InboundData {
+    projectData: AdpProjectData;
+    timestamp: number;
+    /** Identifier for the inbound navigation data. */
+    inboundId: string;
+    flp: {
+        /** Title associated with the inbound navigation data. */
+        title: PropertyValueType;
+        /** Subtitle associated with the inbound navigation data. */
+        subTitle: PropertyValueType;
+        /** Icon associated with the inbound navigation data. */
+        icon: PropertyValueType;
+    };
+    /** Optional flag indicating if the project is in safe mode. */
+    isInSafeMode?: boolean;
+}
+
+export interface InboundContent {
+    inboundId: string;
+    entityPropertyChange: { propertyPath: string; operation: string; propertyValue: unknown }[];
+}
+
+export type DataSourceItem = {
+    uri?: string;
+    type: string;
+    settings: {
+        [key: string]: unknown;
+    };
+};
+
+export type PropertyValueType = 'boolean' | 'number' | 'string' | 'binding' | 'object';
+
+export interface AdpProjectData {
+    path: string;
+    title: string;
+    namespace: string;
+    ui5Version: string;
+    name: string;
+    layer: string;
+    environment: string;
+    safeMode: boolean;
+    sourceSystem: string;
+    applicationIdx: string;
+    reference: string;
+    id: string;
 }