+++ packages/control-property-editor/src/slice.ts
@@ -18,7 +18,9 @@ import {
     propertyChanged,
     propertyChangeFailed,
     showMessage,
-    scenario
+    scenario,
+    reloadApplication,
+    storageFileChanged
 } from '@sap-ux-private/control-property-editor-common';
 import { DeviceType } from './devices';
 
@@ -38,12 +40,14 @@ interface SliceState {
     icons: IconDetails[];
     changes: ChangesSlice;
     dialogMessage: string | undefined;
+    fileChanges?: string[];
 }
 
 export interface ChangesSlice {
     controls: ControlChanges;
     pending: PendingPropertyChange[];
     saved: SavedPropertyChange[];
+    pendingChangeIds: string[];
 }
 export interface ControlChanges {
     [id: string]: ControlChangeStats;
@@ -90,8 +94,9 @@ export const changePreviewScale = createAction<number>('app/change-preview-scale
 export const changePreviewScaleMode = createAction<'fit' | 'fixed'>('app/change-preview-scale-mode');
 export const changeDeviceType = createAction<DeviceType>('app/change-device-type');
 export const filterNodes = createAction<FilterOptions[]>('app/filter-nodes');
-
-export const initialState = {
+export const fileChanged = createAction<string[]>('app/file-changed');
+export const initializeLivereload = createAction<number>('app/initialize-livereload');
+export const initialState: SliceState = {
     deviceType: DeviceType.Desktop,
     scale: 1.0,
     selectedControl: undefined,
@@ -103,7 +108,8 @@ export const initialState = {
     changes: {
         controls: {},
         pending: [],
-        saved: []
+        saved: [],
+        pendingChangeIds: []
     },
     dialogMessage: undefined
 };
@@ -182,6 +188,7 @@ const slice = createSlice<SliceState, SliceCaseReducers<SliceState>, string>({
                 state.changes.saved = action.payload.saved;
                 state.changes.pending = action.payload.pending;
                 state.changes.controls = {};
+
                 for (const change of [...action.payload.pending, ...action.payload.saved].reverse()) {
                     const { controlId, propertyName, type, controlName } = change;
                     const key = `${controlId}`;
@@ -227,6 +234,34 @@ const slice = createSlice<SliceState, SliceCaseReducers<SliceState>, string>({
             })
             .addMatcher(showMessage.match, (state, action: ReturnType<typeof showMessage>): void => {
                 state.dialogMessage = action.payload;
+            })
+            .addMatcher(fileChanged.match, (state, action: ReturnType<typeof fileChanged>): void => {
+                const newFileChanges = action.payload.filter((changedFile) => {
+                    const idx = state.changes.pendingChangeIds.findIndex((pendingFile) =>
+                        changedFile.includes(pendingFile)
+                    );
+                    if (idx > -1) {
+                        state.changes.pendingChangeIds.splice(idx, 1);
+                    }
+                    return idx < 0;
+                });
+                if (!state.fileChanges) {
+                    state.fileChanges = newFileChanges;
+                } else {
+                    state.fileChanges = [
+                        ...state.fileChanges,
+                        ...newFileChanges.filter((changedFile) => !state.fileChanges?.includes(changedFile))
+                    ];
+                }
+            })
+            .addMatcher(reloadApplication.match, (state): void => {
+                state.fileChanges = [];
+            })
+            .addMatcher(storageFileChanged.match, (state, action: ReturnType<typeof storageFileChanged>): void => {
+                const fileName = action.payload;
+                if (fileName) {
+                    state.changes.pendingChangeIds.push(fileName);
+                }
             })
 });