+++ packages/inquirer-common/src/prompts/utility.ts
@@ -1,8 +1,24 @@
-import { Separator, type ListChoiceOptions } from 'inquirer';
+import type { ListChoiceOptions } from 'inquirer';
 import * as fuzzy from 'fuzzy';
 import { t } from '../i18n';
-import type { UI5Version } from '@sap-ux/ui5-info';
+import { getUi5Themes, type UI5Theme, type UI5Version } from '@sap-ux/ui5-info';
 import type { UI5VersionChoice } from '../types';
+import { Separator } from './separator';
+import { coerce, lte, eq } from 'semver';
+
+/**
+ * Get the UI5 themes as prompt choices applicable for the specified UI5 version.
+ *
+ * @param ui5Version - UI5 semantic version
+ * @returns UI5 themes as list choice options
+ */
+export function getUI5ThemesChoices(ui5Version?: string): ListChoiceOptions[] {
+    const themes = getUi5Themes(ui5Version);
+    return themes.map((theme: UI5Theme) => ({
+        name: theme.label,
+        value: theme.id
+    }));
+}
 
 /**
  * Finds the search value in the provided list using `fuzzy` search.
@@ -28,11 +44,13 @@ export function searchChoices(searchVal: string, searchList: ListChoiceOptions[]
  *
  * @param versions ui5Versions
  * @param includeSeparators Include a separator to visually identify groupings, if false then grouping info is included in each entry as additional name text
+ * @param defaultChoice optional, provides an additional version choice entry that is added as the first entry in the version choices and sets as the default
  * @returns Array of ui5 version choices and separators if applicable, grouped by maintenance state
  */
 export function ui5VersionsGrouped(
     versions: UI5Version[],
-    includeSeparators = false
+    includeSeparators = false,
+    defaultChoice?: UI5VersionChoice
 ): (UI5VersionChoice | Separator)[] {
     if (!versions || (Array.isArray(versions) && versions.length === 0)) {
         return [];
@@ -75,5 +93,46 @@ export function ui5VersionsGrouped(
         );
     }
 
-    return [...maintChoices, ...notMaintChoices];
+    const versionChoices = [...maintChoices, ...notMaintChoices];
+    if (defaultChoice) {
+        versionChoices.unshift(defaultChoice);
+    }
+    return versionChoices;
+}
+
+/**
+ * Get the default UI5 version choice that should be selected based on the provided default choice.
+ * Note that if the default choice is not found in the provided versions, the closest provided version is returned.
+ *
+ * @param ui5Versions - UI5 versions ordered by version descending latest first
+ * @param [defaultChoice] - optional default choice to use if found in the provided versions, otherwise the closest provided version is returned
+ * @returns The default UI5 version choice that is closest to the provided default choice or the latest provided version
+ */
+export function getDefaultUI5VersionChoice(
+    ui5Versions: UI5Version[],
+    defaultChoice?: UI5VersionChoice
+): UI5VersionChoice | undefined {
+    if (defaultChoice) {
+        const defaultChoiceVersion = coerce(defaultChoice.value);
+        if (defaultChoiceVersion !== null) {
+            const version = ui5Versions.find((ui5Ver) => lte(ui5Ver.version, defaultChoiceVersion));
+            if (version) {
+                // if the versions are an exact match use the name (UI label) from the default choice as this may use a custom name
+                return {
+                    name: eq(version.version, defaultChoice.value) ? defaultChoice.name : version.version,
+                    value: version.version
+                };
+            }
+        }
+    }
+    // defaultChoice was not coercable, not found or not provided, return the latest version from the ui5 versions provided
+    const defaultVersion = ui5Versions.find((ui5Ver) => ui5Ver.default && ui5Ver.version)?.version;
+
+    if (defaultVersion) {
+        return {
+            name: defaultVersion,
+            value: defaultVersion
+        };
+    }
+    return undefined;
 }