+++ packages/ui-components/test/unit/components/UITable-helper.test.tsx
@@ -1,4 +1,4 @@
-import { waitFor, sleep } from '../../../src/components/UITable/UITable-helper';
+import { waitFor, sleep, _copyAndSort } from '../../../src/components/UITable/UITable-helper';
 
 describe('waitFor', () => {
     it('should resolve with the element when it is found', async () => {
@@ -24,5 +24,43 @@ describe('sleep', () => {
         const end = Date.now();
 
         expect(end - start).toBeGreaterThanOrEqual(100);
+    });
+});
+
+describe('_copyAndSort', () => {
+    const items = [
+        {
+            title: 'b'
+        },
+        {
+            title: 'w'
+        },
+        {
+            title: 'a'
+        },
+        {
+            title: 'ba'
+        }
+    ];
+    const tests = [
+        {
+            isSortedDescending: undefined,
+            expectedOrder: ['w', 'ba', 'b', 'a']
+        },
+        {
+            isSortedDescending: true,
+            expectedOrder: ['a', 'b', 'ba', 'w']
+        },
+        {
+            isSortedDescending: false,
+            expectedOrder: ['w', 'ba', 'b', 'a']
+        }
+    ];
+    test.each(tests)('isSortedDescending = $isSortedDescending', ({ isSortedDescending, expectedOrder }) => {
+        const result = _copyAndSort(items, 'title', isSortedDescending);
+        // Make sure array is copy
+        expect(result).not.toEqual(items);
+        // Check order
+        expect(result.map((item) => item.title)).toEqual(expectedOrder);
     });
 });