+++ packages/adp-tooling/CHANGELOG.md
@@ -1,5 +1,53 @@
 # @sap-ux/adp-tooling
 
+## 0.9.6
+
+### Patch Changes
+
+-   61b46bc8: Security upgrade fixes
+-   Updated dependencies [61b46bc8]
+    -   @sap-ux/axios-extension@1.11.4
+    -   @sap-ux/system-access@0.3.18
+
+## 0.9.5
+
+### Patch Changes
+
+-   Updated dependencies [ec509c40]
+    -   @sap-ux/ui5-config@0.22.0
+    -   @sap-ux/axios-extension@1.11.3
+    -   @sap-ux/system-access@0.3.17
+
+## 0.9.4
+
+### Patch Changes
+
+-   4b29ddcc: Update TypeScript templates, eslint config and ui5 devDependencies
+
+## 0.9.3
+
+### Patch Changes
+
+-   874187b9: Using inbound navigation writer to generate appropriate change files does not produce correct results
+
+## 0.9.2
+
+### Patch Changes
+
+-   212d54ed: Removed unnecessary peer dependency of the adp-tooling and corrected the integration into the preview-middleware
+
+## 0.9.1
+
+### Patch Changes
+
+-   @sap-ux/system-access@0.3.17
+
+## 0.9.0
+
+### Minor Changes
+
+-   cf842794: Adds writer functionality for Adaptation Project's manifest change editors
+
 ## 0.8.11
 
 ### Patch Changes