+++ packages/cds-odata-annotation-converter/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @sap-ux/cds-odata-annotation-converter
 
+## 0.2.4
+
+### Patch Changes
+
+-   Updated dependencies [61b46bc8]
+    -   @sap-ux/odata-vocabularies@0.3.6
+    -   @sap-ux/cds-annotation-parser@0.1.10
+
+## 0.2.3
+
+### Patch Changes
+
+-   Updated dependencies [82c07285]
+    -   @sap-ux/odata-vocabularies@0.3.5
+    -   @sap-ux/cds-annotation-parser@0.1.9
+
 ## 0.2.2
 
 ### Patch Changes