+++ packages/create/CHANGELOG.md
@@ -1,5 +1,132 @@
 # @sap-ux/create
 
+## 0.5.52
+
+### Patch Changes
+
+-   Updated dependencies [87c942e5]
+    -   @sap-ux/project-access@1.19.4
+    -   @sap-ux/adp-tooling@0.9.6
+    -   @sap-ux/app-config-writer@0.3.40
+    -   @sap-ux/cap-config-writer@0.2.24
+    -   @sap-ux/mockserver-config-writer@0.4.0
+
+## 0.5.51
+
+### Patch Changes
+
+-   Updated dependencies [89013210]
+    -   @sap-ux/project-access@1.19.3
+    -   @sap-ux/adp-tooling@0.9.6
+    -   @sap-ux/app-config-writer@0.3.39
+    -   @sap-ux/cap-config-writer@0.2.23
+    -   @sap-ux/mockserver-config-writer@0.4.0
+
+## 0.5.50
+
+### Patch Changes
+
+-   Updated dependencies [6b26a64f]
+    -   @sap-ux/mockserver-config-writer@0.4.0
+
+## 0.5.49
+
+### Patch Changes
+
+-   61b46bc8: Security upgrade fixes
+-   Updated dependencies [61b46bc8]
+    -   @sap-ux/app-config-writer@0.3.38
+    -   @sap-ux/adp-tooling@0.9.6
+
+## 0.5.48
+
+### Patch Changes
+
+-   Updated dependencies [ec509c40]
+    -   @sap-ux/mockserver-config-writer@0.3.0
+    -   @sap-ux/adp-tooling@0.9.5
+    -   @sap-ux/app-config-writer@0.3.37
+    -   @sap-ux/project-access@1.19.2
+    -   @sap-ux/cap-config-writer@0.2.22
+
+## 0.5.47
+
+### Patch Changes
+
+-   @sap-ux/project-access@1.19.1
+-   @sap-ux/adp-tooling@0.9.4
+-   @sap-ux/app-config-writer@0.3.36
+-   @sap-ux/cap-config-writer@0.2.21
+-   @sap-ux/mockserver-config-writer@0.2.3
+
+## 0.5.46
+
+### Patch Changes
+
+-   Updated dependencies [4b29ddcc]
+    -   @sap-ux/adp-tooling@0.9.4
+
+## 0.5.45
+
+### Patch Changes
+
+-   Updated dependencies [a4f00f7c]
+    -   @sap-ux/project-access@1.19.0
+    -   @sap-ux/adp-tooling@0.9.3
+    -   @sap-ux/app-config-writer@0.3.35
+    -   @sap-ux/cap-config-writer@0.2.20
+    -   @sap-ux/mockserver-config-writer@0.2.3
+
+## 0.5.44
+
+### Patch Changes
+
+-   Updated dependencies [874187b9]
+    -   @sap-ux/adp-tooling@0.9.3
+
+## 0.5.43
+
+### Patch Changes
+
+-   Updated dependencies [212d54ed]
+    -   @sap-ux/adp-tooling@0.9.2
+
+## 0.5.42
+
+### Patch Changes
+
+-   Updated dependencies [e6da2117]
+    -   @sap-ux/project-access@1.18.0
+    -   @sap-ux/adp-tooling@0.9.1
+    -   @sap-ux/app-config-writer@0.3.34
+    -   @sap-ux/cap-config-writer@0.2.19
+    -   @sap-ux/mockserver-config-writer@0.2.3
+
+## 0.5.41
+
+### Patch Changes
+
+-   @sap-ux/adp-tooling@0.9.1
+-   @sap-ux/app-config-writer@0.3.33
+
+## 0.5.40
+
+### Patch Changes
+
+-   Updated dependencies [cf842794]
+    -   @sap-ux/adp-tooling@0.9.0
+
+## 0.5.39
+
+### Patch Changes
+
+-   Updated dependencies [c381d32f]
+    -   @sap-ux/project-access@1.17.6
+    -   @sap-ux/adp-tooling@0.8.11
+    -   @sap-ux/app-config-writer@0.3.32
+    -   @sap-ux/cap-config-writer@0.2.18
+    -   @sap-ux/mockserver-config-writer@0.2.3
+
 ## 0.5.38
 
 ### Patch Changes