+++ packages/deploy-tooling/CHANGELOG.md
@@ -1,5 +1,43 @@
 # @sap-ux/deploy-tooling
 
+## 0.14.8
+
+### Patch Changes
+
+-   4c7d71eb: Adjust logging in validation for deploy-test to not log full stack trace, but only the message.
+
+## 0.14.7
+
+### Patch Changes
+
+-   61b46bc8: Security upgrade fixes
+-   Updated dependencies [61b46bc8]
+    -   @sap-ux/axios-extension@1.11.4
+    -   @sap-ux/btp-utils@0.14.3
+    -   @sap-ux/system-access@0.3.18
+
+## 0.14.6
+
+### Patch Changes
+
+-   Updated dependencies [ec509c40]
+    -   @sap-ux/ui5-config@0.22.0
+    -   @sap-ux/axios-extension@1.11.3
+    -   @sap-ux/system-access@0.3.17
+
+## 0.14.5
+
+### Patch Changes
+
+-   @sap-ux/system-access@0.3.17
+
+## 0.14.4
+
+### Patch Changes
+
+-   Updated dependencies [1affcec6]
+    -   @sap-ux/project-input-validator@0.2.3
+
 ## 0.14.3
 
 ### Patch Changes