+++ packages/fiori-elements-writer/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @sap-ux/fiori-elements-writer
 
+## 0.18.10
+
+### Patch Changes
+
+-   Updated dependencies [ec509c40]
+    -   @sap-ux/ui5-config@0.22.0
+    -   @sap-ux/odata-service-writer@0.16.2
+    -   @sap-ux/ui5-application-writer@0.26.5
+    -   @sap-ux/fe-fpm-writer@0.24.9
+    -   @sap-ux/ui5-test-writer@0.3.2
+
+## 0.18.9
+
+### Patch Changes
+
+-   4b29ddcc: Update TypeScript templates, eslint config and ui5 devDependencies
+-   Updated dependencies [4b29ddcc]
+    -   @sap-ux/ui5-application-writer@0.26.4
+
+## 0.18.8
+
+### Patch Changes
+
+-   7964a7ba: Adds exports for determining supported ui5 versions for FE templates
+
 ## 0.18.7
 
 ### Patch Changes