+++ packages/telemetry/CHANGELOG.md
@@ -1,5 +1,87 @@
 # @sap-ux/telemetry
 
+## 0.4.10
+
+### Patch Changes
+
+-   Updated dependencies [87c942e5]
+    -   @sap-ux/project-access@1.19.4
+
+## 0.4.9
+
+### Patch Changes
+
+-   Updated dependencies [89013210]
+    -   @sap-ux/project-access@1.19.3
+
+## 0.4.8
+
+### Patch Changes
+
+-   61b46bc8: Security upgrade fixes
+-   Updated dependencies [61b46bc8]
+    -   @sap-ux/btp-utils@0.14.3
+
+## 0.4.7
+
+### Patch Changes
+
+-   Updated dependencies [ec509c40]
+    -   @sap-ux/ui5-config@0.22.0
+    -   @sap-ux/project-access@1.19.2
+
+## 0.4.6
+
+### Patch Changes
+
+-   @sap-ux/project-access@1.19.1
+
+## 0.4.5
+
+### Patch Changes
+
+-   Updated dependencies [a4f00f7c]
+    -   @sap-ux/project-access@1.19.0
+
+## 0.4.4
+
+### Patch Changes
+
+-   e05c129b: Remove unused dependency
+
+## 0.4.3
+
+### Patch Changes
+
+-   Updated dependencies [e6da2117]
+    -   @sap-ux/project-access@1.18.0
+
+## 0.4.2
+
+### Patch Changes
+
+-   Updated dependencies [e443e534]
+    -   @sap-ux/store@0.5.0
+
+## 0.4.1
+
+### Patch Changes
+
+-   Updated dependencies [c381d32f]
+    -   @sap-ux/project-access@1.17.6
+
+## 0.4.0
+
+### Minor Changes
+
+-   15d6afbe: update init settings, bump azure applicationinsights and stablise tests
+
+## 0.3.1
+
+### Patch Changes
+
+-   98d34e8e: fix for assigning resource id in telemetry settings
+
 ## 0.3.0
 
 ### Minor Changes